examples/strands-agent/pyproject.toml
[project]
name = "strands-agent-example"
version = "0.1.0"
description = "A simple Strands agent for AWS AgentCore runtime"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.6.0",
    "strands-agents>=1.34.1",
    "strands-agents-tools>=0.3.0",
    "awscrt>=0.32.0",
]

[project.scripts]
strands-agent = "strands_agent_example.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strands_agent_example"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

